webpack

+++ src/controllers/videoController.js
@@ -8,25 +8,32 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
     const { id } = req.params;
     const video = await Video.findById(id).populate("owner");
-    console.log(video);
     if (!video)
         return res.render("404", { pageTitle: "Video not found." });
     return res.render("watch", { pageTitle: video.title, video });
 }
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const video = await Video.findById(id);
     if (!video)
         return res.render("404", { pageTitle: "Video not found." });
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    };
     return res.render("edit", { pageTitle: `Editing`, video });
 }
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    };
     await Video.findByIdAndUpdate(id, {
         title, description,
         hashtags: Video.formatHashtags(hashtags)
@@ -62,6 +69,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    };
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
